servidor/requirements.txt
Flask
flask-cors
flask-jwt-extended
mysql-connector-python
Pillow
PyMuPDF
python-dotenv
